Move offsets-tracker to internal/tools pkg
