+++ internal/tools/offsets/.gitignore
