+++ internal/tools/offsets/Dockerfile
@@ -1,0 +1,5 @@
+FROM golang:1.20
+COPY . /offsets
+WORKDIR /offsets
+RUN go build -o offsets
+ENTRYPOINT ["/offsets/offsets"]
