+++ internal/tools/offsets/LICENSE
