+++ internal/tools/offsets/binary/data_member.go
