+++ internal/tools/offsets/binary/reader.go
