+++ internal/tools/offsets/downloader/go_wrapper.go
@@ -21,7 +21,7 @@ import (
 	"os"
 	"path"
 
-	"go.opentelemetry.io/auto/offsets-tracker/utils"
+	"go.opentelemetry.io/auto/internal/tools/offsets/utils"
 )
 
 const appName = "testapp"
