+++ internal/tools/offsets/downloader/remote_targz.go
@@ -22,7 +22,7 @@ import (
 	"os"
 	"path"
 
-	"go.opentelemetry.io/auto/offsets-tracker/utils"
+	"go.opentelemetry.io/auto/internal/tools/offsets/utils"
 )
 
 const (
