+++ internal/tools/offsets/downloader/wrapper/go.mod.txt
