+++ internal/tools/offsets/main.go
@@ -22,9 +22,9 @@ import (
 
 	"github.com/hashicorp/go-version"
 
-	"go.opentelemetry.io/auto/offsets-tracker/binary"
-	"go.opentelemetry.io/auto/offsets-tracker/target"
-	"go.opentelemetry.io/auto/offsets-tracker/writer"
+	"go.opentelemetry.io/auto/internal/tools/offsets/binary"
+	"go.opentelemetry.io/auto/internal/tools/offsets/target"
+	"go.opentelemetry.io/auto/internal/tools/offsets/writer"
 )
 
 const (
