+++ internal/tools/offsets/schema/schema.go
