+++ internal/tools/offsets/utils/cmd.go
