+++ internal/tools/offsets/versions/golang_org.go
