--- offsets-tracker/go.mod
@@ -1,5 +1,0 @@
-module go.opentelemetry.io/auto/offsets-tracker
-
-go 1.17
-
-require github.com/hashicorp/go-version v1.6.0
